chore: rename build command dir
