+++ src/commands/deploy/prompts/getProjectPathOrPrompt.ts
